Move client object to client module #172
