+++ dams-client/src/api/authenticate.rs
@@ -1,3 +1,6 @@
+use crate::client::{DamsClient, Password};
+
+use crate::DamsClientError;
 use dams::{
     channel::ClientChannel,
     config::opaque::OpaqueCipherSuite,
@@ -9,32 +12,32 @@ use opaque_ke::{
 };
 use rand::{CryptoRng, RngCore};
 
-use crate::{api::Password, error::DamsClientError};
+impl DamsClient {
+    pub(crate) async fn handle_authentication<T: CryptoRng + RngCore>(
+        mut channel: ClientChannel,
+        rng: &mut T,
+        user_id: &UserId,
+        password: &Password,
+    ) -> Result<[u8; 64], DamsClientError> {
+        let client_login_start_result =
+            ClientLogin::<OpaqueCipherSuite>::start(rng, password.as_bytes())?;
 
-pub(crate) async fn handle<T: CryptoRng + RngCore>(
-    mut channel: ClientChannel,
-    rng: &mut T,
-    user_id: &UserId,
-    password: &Password,
-) -> Result<[u8; 64], DamsClientError> {
-    let client_login_start_result =
-        ClientLogin::<OpaqueCipherSuite>::start(rng, password.as_bytes())?;
+        // Handle start step
+        let server_start_result =
+            authenticate_start(&mut channel, &client_login_start_result, user_id).await?;
 
-    // Handle start step
-    let server_start_result =
-        authenticate_start(&mut channel, &client_login_start_result, user_id).await?;
+        // Handle finish step
+        let client_login_finish_result = authenticate_finish(
+            &mut channel,
+            user_id,
+            password,
+            client_login_start_result,
+            server_start_result,
+        )
+        .await?;
 
-    // Handle finish step
-    let client_login_finish_result = authenticate_finish(
-        &mut channel,
-        user_id,
-        password,
-        client_login_start_result,
-        server_start_result,
-    )
-    .await?;
-
-    Ok(client_login_finish_result.session_key.into())
+        Ok(client_login_finish_result.session_key.into())
+    }
 }
 
 async fn authenticate_start(
